pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
streamlit = "^1.27.2"
langchain = "^0.1.17"
langchain-core = "^0.1.50"
langchain-google-genai = "^1.0.3"
langchain-openai = "^0.1.6"
pandas = "^2.2.2"
seaborn = "^0.13.2"
matplotlib = "^3.8.4"
openai = "^1.30.1"
open-interpreter = "^0.2.5"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"